Cast bytes to int before multiplying it.
Removes warning in PHP 7.0+

+++ h5p.classes.php
@@ -2969,7 +2969,7 @@ class H5PCore {
    * @return int|string
    */
   public static function returnBytes($val) {
-    $val  = trim($val);
+    $val  = (int) trim($val);
     $last = strtolower($val[strlen($val) - 1]);
     switch ($last) {
       case 'g':
